# Update to Serilog 4.x, remove some allocations