--- a/src/Serilog.Extensions.Logging/Extensions/Logging/SerilogLogger.cs
+++ b/src/Serilog.Extensions.Logging/Extensions/Logging/SerilogLogger.cs
@@ -103,12 +103,12 @@ class SerilogLogger : FrameworkLogger
                 {
                     messageTemplate = value;
                 }
-                else if (property.Key.StartsWith("@"))
+                else if (property.Key.StartsWith("@", StringComparison.Ordinal))
                 {
                     if (_logger.BindProperty(GetKeyWithoutFirstSymbol(DestructureDictionary, property.Key), property.Value, true, out var destructured))
                         properties.Add(destructured.Name, destructured.Value);
                 }
-                else if (property.Key.StartsWith("$"))
+                else if (property.Key.StartsWith("$", StringComparison.Ordinal))
                 {
                     if (_logger.BindProperty(GetKeyWithoutFirstSymbol(StringifyDictionary, property.Key), property.Value?.ToString(), true, out var stringified))
                         properties.Add(stringified.Name, stringified.Value);